Add/update git decor colors, dull blue slightly more than red and green
Even in LCH / CIELAB, blues tend to look more saturated, which was pretty
noticeable when I was editing the git decor colors, so I decreased chroma a
little more than red and green.
Added dull versions of red, green, blue, and orange for the git decor colors so
that they don't stand out over the primary color of each theme.